refactor: remove barrel file and update imports
- Delete token.transaction.schema.ts barrel file
- Update all affected imports to use direct import path
- Follows project standard of avoiding barrel files (index.ts exports)
