--- kit/dapp/src/orpc/routes/token/routes/mutations/common/token.transaction.schema.ts
@@ -1,7 +1,0 @@
-import { MutationOutputSchema } from "@/orpc/routes/common/schemas/mutation.schema";
-
-/**
- * Generic output schema for token transaction operations
- * This is used for all token operations that return a transaction hash
- */
-export const TokenTransactionOutputSchema = MutationOutputSchema;
