+++ kit/dapp/src/orpc/routes/token/routes/mutations/compliance/token.remove-compliance-module.contract.ts
@@ -1,5 +1,5 @@
 import { baseContract } from "@/orpc/procedures/base.contract";
-import { TokenTransactionOutputSchema } from "@/orpc/routes/token/routes/mutations/common/token.transaction.schema";
+import { MutationOutputSchema as TokenTransactionOutputSchema } from "@/orpc/routes/common/schemas/mutation.schema";
 import { TokenRemoveComplianceModuleInputSchema } from "@/orpc/routes/token/routes/mutations/compliance/token.remove-compliance-module.schema";
 import { eventIterator } from "@orpc/server";
