+++ kit/dapp/src/orpc/routes/token/routes/mutations/recovery/token.recover-erc20.contract.ts
@@ -1,6 +1,6 @@
 import { baseContract } from "@/orpc/procedures/base.contract";
 import { TokenRecoverERC20InputSchema } from "@/orpc/routes/token/routes/mutations/recovery/token.recover-erc20.schema";
-import { TokenTransactionOutputSchema } from "@/orpc/routes/token/routes/mutations/common/token.transaction.schema";
+import { MutationOutputSchema as TokenTransactionOutputSchema } from "@/orpc/routes/common/schemas/mutation.schema";
 import { eventIterator } from "@orpc/server";
 
 export const tokenRecoverERC20Contract = baseContract
